v0.12.12. Many changes to the way data reading and addresses are handled. It should work correctly when SC2 is closed and reopened.

+++ 2cs-API_Source/_2cs_API/_2csAPI.cs
@@ -1904,7 +1904,7 @@ namespace _2cs_API
 				{
 					try
 					{
-						return GameData.GetPlayer(LocalPlayerNumber);
+						return GameData.GetPlayer((int)LocalPlayerNumber);
 					}
 					catch
 					{
